refactor: package structure and names
