+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = tralda
-version = 1.1.1
+version = 2.0.0
 author = David Schaller
 author_email = sdavid@bioinf.uni-leipzig.de
 description = A library for tree data structures and algorithms
