+++ src/tralda/cograph/__init__.py
@@ -1,12 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from tralda.cograph.Cograph import (to_cograph,
-                                    to_cotree,
-                                    complement_cograph,
-                                    random_cotree,
-                                    LinearCographDetector)
+from tralda.cograph.functions import (to_cograph,
+                                      to_cotree,
+                                      complement_cograph,
+                                      random_cotree,
+                                      cluster_deletion,
+                                      complete_multipartite_completion,)
 
-from tralda.cograph.CographEditor import edit_to_cograph, CographEditor
-
-from tralda.cograph.ClusterDeletion import (cluster_deletion,
-                                            complete_multipartite_completion)
+from tralda.cograph.editing import edit_to_cograph, CographEditor
