+++ src/tralda/datastructures/Partition.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from tralda.datastructures.DoublyLinkedList import DLList
+from tralda.datastructures.doubly_linked import DLList
 
 
 __author__ = 'David Schaller'
