+++ src/tralda/datastructures/bst/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+"""
+Binary search trees subpackage.
+"""
