+++ src/tralda/datastructures/bst/base.py
