+++ src/tralda/datastructures/doubly_linked.py
