+++ src/tralda/datastructures/linked.py
