+++ src/tralda/supertree/__init__.py
@@ -1,17 +1,17 @@
 # -*- coding: utf-8 -*-
 
-from tralda.supertree.Build import (Build,
+from tralda.supertree.build import (Build,
                                     BUILD_supertree,
                                     tree_profile_to_triples,
                                     greedy_BUILD,
                                     best_pair_merge_first)
 
-from tralda.supertree.BuildST import (BuildST,
+from tralda.supertree.build_st import (BuildST,
                                       build_st)
 
-from tralda.supertree.CommonRefinement import (LinCR,
+from tralda.supertree.common_refinement import (LinCR,
                                                linear_common_refinement)
 
-from tralda.supertree.LooseConsensus import (loose_consensus_tree,
-                                             LooseConsensusTree,
-                                             merge_all)
+from tralda.supertree.consensus import (loose_consensus_tree,
+                                        LooseConsensusTree,
+                                        merge_all)
